introduce laravel breeze

+++ app/Http/Controllers/TasksController.php
@@ -15,8 +15,15 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
-
+        $tasks = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザの投稿の一覧を作成日時の降順で取得
+            // （後のChapterで他ユーザの投稿も取得するように変更しますが、現時点ではこのユーザの投稿のみ取得します）
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+        }
+        
         return view('tasks.index', [
             'tasks' => $tasks,
         ]);
@@ -50,11 +57,12 @@ class TasksController extends Controller
             'content' => 'required',
         ]);
         
-        $task = new Task;
-        $task->status = $request->status;    // 追加
-        $task->content = $request->content;
-        $task->save();
-
+        // 認証済みユーザ（閲覧者）の投稿として作成（リクエストされた値をもとに作成）
+        $request->user()->tasks()->create([
+            'status' =>$request->status,
+            'content' => $request->content,
+        ]);
+        
         return redirect('/');
     }
 
@@ -66,8 +74,9 @@ class TasksController extends Controller
      */
     public function show($id)
     {
+        // idの値でメッセージを検索して取得
         $task = Task::findOrFail($id);
-
+        
         return view('tasks.show', [
             'task' => $task,
         ]);
@@ -120,9 +129,14 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        $task = Task::findOrFail($id);
-        $task->delete();
-
-        return redirect('/');
+        // idの値で投稿を検索して取得
+        $task = \App\Models\Task::findOrFail($id);
+        
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+            return redirect('/')
+                ->with('success','Delete Successful');
+        }
     }
 }
